requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-digitalocean>=4.0.0,<5.0.0
pyyaml>=6.0.0,<7.0.0
